backend/gpubroker/apps/math_core/api.py
"""
Math Core API - Django Ninja Router.

Endpoints for mathematical calculations, TOPSIS, and workload mapping.
NO MOCKS. NO FAKE DATA. REAL IMPLEMENTATIONS ONLY.
"""
import logging

import numpy as np
from ninja import Router
from ninja.errors import HttpError

from .algorithms.topsis import TOPSISEngine
from .benchmarks import benchmark_repo
from .schemas import (
    BenchmarkListResponse,
    CostPerGFLOPRequest,
    CostPerGFLOPResponse,
    CostPerTokenRequest,
    CostPerTokenResponse,
    EfficiencyScoreRequest,
    EfficiencyScoreResponse,
    EnsembleRecommendRequest,
    EnsembleRecommendResponse,
    GPUBenchmarkSchema,
    TOPSISRequest,
    TOPSISResponse,
    WorkloadEstimateRequest,
    WorkloadEstimateResponse,
)
from .services import kpi_calculator, workload_mapper

logger = logging.getLogger('gpubroker.math_core.api')

router = Router()

# Initialize TOPSIS engine
topsis_engine = TOPSISEngine()


# ============================================
# KPI Calculation Endpoints
# ============================================

@router.post("/cost-per-token", response=CostPerTokenResponse)
def calculate_cost_per_token(request, payload: CostPerTokenRequest):
    """
    Calculate cost per token for LLM inference workloads.
    
    Uses IEEE 754 double precision with 4 decimal rounding.
    """
    try:
        result = kpi_calculator.cost_per_token(
            price_per_hour=payload.price_per_hour,
            gpu_type=payload.gpu_type,
            model_size=payload.model_size
        )
        return result
    except ValueError as e:
        raise HttpError(400, str(e))
    except Exception as e:
        logger.exception(f"Cost per token calculation failed: {e}")
        raise HttpError(500, "Calculation failed")


@router.post("/cost-per-gflop", response=CostPerGFLOPResponse)
def calculate_cost_per_gflop(request, payload: CostPerGFLOPRequest):
    """
    Calculate cost per GFLOP for compute workloads.
    """
    try:
        result = kpi_calculator.cost_per_gflop(
            price_per_hour=payload.price_per_hour,
            gpu_type=payload.gpu_type,
            precision=payload.precision
        )
        return result
    except ValueError as e:
        raise HttpError(400, str(e))
    except Exception as e:
        logger.exception(f"Cost per GFLOP calculation failed: {e}")
        raise HttpError(500, "Calculation failed")


@router.post("/efficiency-score", response=EfficiencyScoreResponse)
def calculate_efficiency_score(request, payload: EfficiencyScoreRequest):
    """
    Calculate overall efficiency score combining price, performance, and reliability.
    """
    try:
        result = kpi_calculator.efficiency_score(
            price_per_hour=payload.price_per_hour,
            gpu_type=payload.gpu_type,
            availability_score=payload.availability_score,
            reliability_score=payload.reliability_score
        )
        return result
    except Exception as e:
        logger.exception(f"Efficiency score calculation failed: {e}")
        raise HttpError(500, "Calculation failed")


# ============================================
# TOPSIS Endpoint
# ============================================

@router.post("/topsis", response=TOPSISResponse)
def run_topsis(request, payload: TOPSISRequest):
    """
    Execute TOPSIS (Technique for Order of Preference by Similarity to Ideal Solution).
    
    Multi-criteria decision analysis algorithm for ranking alternatives.
    """
    try:
        # Validate weights sum to 1
        weight_sum = sum(payload.weights)
        if abs(weight_sum - 1.0) > 0.001:
            raise HttpError(400, f"Weights must sum to 1.0, got {weight_sum}")
        
        result = topsis_engine.calculate(
            decision_matrix=np.array(payload.decision_matrix),
            weights=payload.weights,
            criteria_types=payload.criteria_types
        )
        return result
    except ValueError as e:
        raise HttpError(400, str(e))
    except HttpError:
        raise
    except Exception as e:
        logger.exception(f"TOPSIS calculation failed: {e}")
        raise HttpError(500, "TOPSIS calculation failed")


# ============================================
# Ensemble Recommendations Endpoint
# ============================================

@router.post("/ensemble-recommend", response=EnsembleRecommendResponse)
def ensemble_recommend(request, payload: EnsembleRecommendRequest):
    """
    Generate recommendations using weighted ensemble of algorithms.
    
    Combines TOPSIS (0.4) + Content-Based (0.6) for cold-start scenarios.
    Full collaborative filtering requires user history.
    """
    try:
        if not payload.candidate_offers:
            return {
                "recommendations": [],
                "algorithm_contributions": {"topsis": 0.0, "content_based": 0.0},
                "confidence": 0.0
            }
        
        # Use TOPSIS for ranking
        criteria_config = {
            "price_per_hour": {"weight": 0.30, "type": "cost"},
            "vram_gb": {"weight": 0.25, "type": "benefit"},
            "tflops": {"weight": 0.20, "type": "benefit"},
            "availability_score": {"weight": 0.15, "type": "benefit"},
            "reliability_score": {"weight": 0.10, "type": "benefit"}
        }
        
        # Adjust weights based on workload profile
        budget = payload.workload_profile.get("budget_per_hour")
        if budget and budget < 1.0:
            criteria_config["price_per_hour"]["weight"] = 0.40
            criteria_config["vram_gb"]["weight"] = 0.20
        
        # Build decision matrix
        criteria_names = list(criteria_config.keys())
        n_offers = len(payload.candidate_offers)
        n_criteria = len(criteria_names)
        
        matrix = np.zeros((n_offers, n_criteria))
        for i, offer in enumerate(payload.candidate_offers):
            for j, criterion in enumerate(criteria_names):
                value = offer.get(criterion, 0)
                if value is None:
                    defaults = {
                        "availability_score": 0.8,
                        "reliability_score": 0.8,
                        "tflops": 20.0,
                        "vram_gb": 16
                    }
                    value = defaults.get(criterion, 0)
                matrix[i, j] = float(value)
        
        weights = [criteria_config[c]["weight"] for c in criteria_names]
        criteria_types = [criteria_config[c]["type"] for c in criteria_names]
        
        topsis_result = topsis_engine.calculate(matrix, weights, criteria_types)
        scores = np.array(topsis_result["scores"])
        
        # Normalize scores
        min_score = np.min(scores)
        max_score = np.max(scores)
        if max_score - min_score > 1e-10:
            normalized_scores = (scores - min_score) / (max_score - min_score)
        else:
            normalized_scores = np.ones_like(scores) * 0.5
        
        # Rank and select top-k
        ranked_indices = np.argsort(-normalized_scores)[:payload.top_k]
        
        recommendations = []
        for idx in ranked_indices:
            offer = payload.candidate_offers[idx].copy()
            offer["ensemble_score"] = float(normalized_scores[idx])
            offer["score_breakdown"] = {
                "topsis": float(scores[idx]),
                "content_based": 0.0  # Simplified for now
            }
            recommendations.append(offer)
        
        return {
            "recommendations": recommendations,
            "algorithm_contributions": {"topsis": 1.0, "content_based": 0.0},
            "confidence": 0.85
        }
    except Exception as e:
        logger.exception(f"Ensemble recommendation failed: {e}")
        raise HttpError(500, "Recommendation failed")


# ============================================
# Workload Estimation Endpoint
# ============================================

@router.post("/estimate-workload", response=WorkloadEstimateResponse)
def estimate_workload(request, payload: WorkloadEstimateRequest):
    """
    Estimate GPU requirements for a given workload.
    
    Translates workload descriptions to GPU specs.
    """
    try:
        result = workload_mapper.estimate(
            workload_type=payload.workload_type,
            quantity=payload.quantity,
            duration_hours=payload.duration_hours,
            quality=payload.quality,
            model_name=payload.model_name
        )
        return result
    except ValueError as e:
        raise HttpError(400, str(e))
    except Exception as e:
        logger.exception(f"Workload estimation failed: {e}")
        raise HttpError(500, "Workload estimation failed")


# ============================================
# Benchmark Endpoints
# ============================================

@router.get("/benchmarks", response=BenchmarkListResponse)
def list_benchmarks(request):
    """
    List all GPU benchmarks in the database.
    """
    benchmarks = benchmark_repo.list_all()
    return {
        "benchmarks": benchmarks,
        "count": len(benchmarks)
    }


@router.get("/benchmarks/{gpu_type}", response=GPUBenchmarkSchema)
def get_benchmark(request, gpu_type: str):
    """
    Get benchmark data for a specific GPU type.
    """
    benchmark = benchmark_repo.get(gpu_type)
    if not benchmark:
        raise HttpError(404, f"Benchmark not found for GPU: {gpu_type}")
    
    return {
        "gpu_model": gpu_type,
        **benchmark
    }
