pyproject.toml
[tool.poetry]
name = "gpubroker"
version = "0.1.0"
description = "GPUBROKER control‑tower SaaS"
authors = ["SomaAgent 01 <somaagent01@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^5.0"
django-ninja = "^1.1"
channels = "^4.0"
channels-redis = "^4.2"
daphne = "^4.1"
psycopg2-binary = "^2.9"
redis = "^5.0"
django-redis = "^5.4"
clickhouse-driver = "^0.2"
langchain = "^0.1.0"
pymoo = "^0.6"
xgboost = "^2.0"
pydantic = "^2.6"
python-jose = "^3.3"
passlib = {extras = ["argon2"], version = "^1.7"}
django-prometheus = "^2.3"
opentelemetry-instrumentation-django = "^0.44b0"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
ruff = "^0.2"
